+++ NEWS.md
@@ -1,5 +1,10 @@
 # teal.picks 0.2.0.9002
 
+### Enhancements
+
+* Fixed `variables` selection popup to show "Select all"/"Deselect all" when multiple selection is enabled.
+* Adds assertion to check for specific elements of picks to API.
+
 # teal.picks 0.2.0
 
 ### Bug fixes